+++ pkg/cluster/internal/create/actions/createworker/provider.go
@@ -24,6 +24,8 @@ import (
 	"fmt"
 	"io/ioutil"
 	"path/filepath"
+	"regexp"
+	"strconv"
 	"time"
 
 	"strings"
@@ -44,6 +46,9 @@ var denyAllEgressIMDSgnpFiles embed.FS
 
 //go:embed files/*/allow-egress-imds_gnetpol.yaml
 var allowEgressIMDSgnpFiles embed.FS
+
+//go:embed files/*/*_pdb.yaml
+var capxPDBFile embed.FS
 
 const (
 	CAPICoreProvider         = "cluster-api:v1.5.1"
@@ -224,6 +229,21 @@ func (p *Provider) getAllowCAPXEgressIMDSGNetPol() (string, error) {
 	}
 
 	return string(allowEgressIMDSgnpContent), nil
+}
+
+func (p *Provider) getcapxPDB() (string, error) {
+	capxPDBLocalPath := "files/" + p.capxProvider + "/" + p.capxName + "_pdb.yaml"
+	capxPDBFile, err := capxPDBFile.Open(capxPDBLocalPath)
+	if err != nil {
+		return "", errors.Wrap(err, "error opening the PodDisruptionBudget file")
+	}
+	defer capxPDBFile.Close()
+	capaPDBContent, err := ioutil.ReadAll(capxPDBFile)
+	if err != nil {
+		return "", err
+	}
+
+	return string(capaPDBContent), nil
 }
 
 func deployClusterOperator(n nodes.Node, keosCluster commons.KeosCluster, clusterCredentials commons.ClusterCredentials, keosRegistry keosRegistry, kubeconfigPath string, firstInstallation bool) error {
@@ -448,6 +468,8 @@ func (p *Provider) installCAPXWorker(n nodes.Node, kubeconfigPath string, allowA
 	var c string
 	var err error
 
+	capxPDBPath := "/kind/" + p.capxName + "_pdb.yaml"
+
 	if p.capxProvider == "azure" {
 		// Create capx namespace
 		c = "kubectl --kubeconfig " + kubeconfigPath + " create namespace " + p.capxName + "-system"
@@ -476,6 +498,93 @@ func (p *Provider) installCAPXWorker(n nodes.Node, kubeconfigPath string, allowA
 		return errors.Wrap(err, "failed to install CAPX in workload cluster")
 	}
 
+	// Manually assign PriorityClass to capa service
+	c = "kubectl --kubeconfig " + kubeconfigPath + " -n " + p.capxName + "-system patch deploy " + p.capxName + "-controller-manager -p '{\"spec\": {\"template\": {\"spec\": {\"priorityClassName\": \"system-node-critical\"}}}}' --type=merge"
+	_, err = commons.ExecuteCommand(n, c)
+	if err != nil {
+		return errors.Wrap(err, "failed to assigned priorityClass to " + p.capxName +"-controller-manager")
+	}
+	c = "kubectl --kubeconfig " + kubeconfigPath + " -n " + p.capxName + "-system rollout status deploy " + p.capxName + "-controller-manager --timeout 60s"
+	if err != nil {
+		return errors.Wrap(err, "failed to check rollout status for " + p.capxName +"-controller-manager")
+	}
+
+	// Scale CAPX to 2 replicas
+	c = "kubectl --kubeconfig " + kubeconfigPath + " -n " + p.capxName + "-system scale --replicas 2 deploy " + p.capxName + "-controller-manager"
+	_, err = commons.ExecuteCommand(n, c)
+	if err != nil {
+		return errors.Wrap(err, "failed to scale CAPX in workload cluster")
+	}
+	c = "kubectl --kubeconfig " + kubeconfigPath + " -n " + p.capxName + "-system rollout status deploy " + p.capxName + "-controller-manager --timeout 60s"
+	if err != nil {
+		return errors.Wrap(err, "failed to check rollout status for " + p.capxName +"-controller-manager")
+	}
+
+	// Define PodDisruptionBudget for capa service
+
+	capxPDB, err := p.getcapxPDB()
+	if err != nil {
+		return err
+	}
+
+	c = "echo \"" + capxPDB + "\" > " + capxPDBPath
+	_, err = commons.ExecuteCommand(n, c)
+	if err != nil {
+		return errors.Wrap(err, "failed to create "+p.capxName+"PodDisruptionBudget file")
+	}
+
+	c = "kubectl --kubeconfig " + kubeconfigPath + " apply -f " + capxPDBPath
+	_, err = commons.ExecuteCommand(n, c)
+	if err != nil {
+		return errors.Wrap(err, "failed to apply "+p.capxName+" PodDisruptionBudget")
+	}
+
+	// Allow egress in CAPX's Namespace
+	c = "kubectl --kubeconfig " + kubeconfigPath + " -n " + p.capxName + "-system apply -f " + allowAllEgressNetPolPath
+	_, err = commons.ExecuteCommand(n, c)
+	if err != nil {
+		return errors.Wrap(err, "failed to apply CAPX's NetworkPolicy in workload cluster")
+	}
+
+	return nil
+}
+
+func (p *Provider) configCAPIWorker(n nodes.Node, keosCluster commons.KeosCluster, kubeconfigPath string, allowCommonEgressNetPolPath string) error {
+	var c string
+	var err error
+	var capiKubeadmReplicas int
+
+	capiDeployments := []struct {
+		name      string
+		namespace string
+	}{
+		{name: "capi-controller-manager", namespace: "capi-system"},
+		{name: "capi-kubeadm-control-plane-controller-manager", namespace: "capi-kubeadm-control-plane-system"},
+		{name: "capi-kubeadm-bootstrap-controller-manager", namespace: "capi-kubeadm-bootstrap-system"},
+	}
+
+	allowedNamePattern := regexp.MustCompile(`^capi-kubeadm-(control-plane|bootstrap)-controller-manager$`)
+	capiPDBPath := "/kind/capi_pdb.yaml"
+
+	// Determine the number of replicas for capi-kubeadm deployments
+	if p.capxManaged {
+		capiKubeadmReplicas = 0
+	} else {
+		capiKubeadmReplicas = 2
+	}
+
+	// Manually assign PriorityClass to capi services
+	for _, deployment := range capiDeployments {
+		if !p.capxManaged || (p.capxManaged && !allowedNamePattern.MatchString(deployment.name)) {
+			c = "kubectl --kubeconfig " + kubeconfigPath + " -n " + deployment.namespace + " patch deploy " + deployment.name + " -p '{\"spec\": {\"template\": {\"spec\": {\"priorityClassName\": \"system-node-critical\"}}}}' --type=merge"
+			_, err = commons.ExecuteCommand(n, c)
+			if err != nil {
+				return errors.Wrap(err, "failed to assigned priorityClass to "+deployment.name)
+			}
+		}
+	}
+
+	// Manually assign PriorityClass to nmi
 	if p.capxProvider == "azure" {
 		c = "kubectl --kubeconfig " + kubeconfigPath + " -n " + p.capxName + "-system patch ds capz-nmi -p '{\"spec\": {\"template\": {\"spec\": {\"priorityClassName\": \"system-node-critical\"}}}}' --type=merge"
 		_, err = commons.ExecuteCommand(n, c)
@@ -488,18 +597,65 @@ func (p *Provider) installCAPXWorker(n nodes.Node, kubeconfigPath string, allowA
 		}
 	}
 
-	// Scale CAPX to 2 replicas
-	c = "kubectl --kubeconfig " + kubeconfigPath + " -n " + p.capxName + "-system scale --replicas 2 deploy " + p.capxName + "-controller-manager"
+	// Scale number of replicas to 2 for capi service
+	c = "kubectl --kubeconfig " + kubeconfigPath + " -n capi-system scale deploy capi-controller-manager --replicas 2"
 	_, err = commons.ExecuteCommand(n, c)
 	if err != nil {
-		return errors.Wrap(err, "failed to scale CAPX in workload cluster")
+		return errors.Wrap(err, "failed to scale the CAPI Deployment")
+	}
+	c = "kubectl --kubeconfig " + kubeconfigPath + " -n capi-system rollout status deploy capi-controller-manager --timeout 60s"
+	if err != nil {
+		return errors.Wrap(err, "failed to check rollout status for capi-controller-manager")
 	}
 
-	// Allow egress in CAPX's Namespace
-	c = "kubectl --kubeconfig " + kubeconfigPath + " -n " + p.capxName + "-system apply -f " + allowAllEgressNetPolPath
+	// Scale number of required replicas for capi kubeadm services
+	for _, deployment := range capiDeployments {
+		if deployment.name != "capi-controller-manager" {
+			c = "kubectl --kubeconfig " + kubeconfigPath + " -n " + deployment.namespace + " scale --replicas " + strconv.Itoa(capiKubeadmReplicas) + " deploy " + deployment.name
+			_, err = commons.ExecuteCommand(n, c)
+			if err != nil {
+				return errors.Wrap(err, "failed to scale the "+deployment.name+" deployment")
+			}
+			c = "kubectl --kubeconfig " + kubeconfigPath + " -n capi-system rollout status deploy " + deployment.name + " --timeout 60s"
+			if err != nil {
+				return errors.Wrap(err, "failed to check rollout status for "+deployment.name)
+			}
+		}
+	}
+
+	// Define PodDisruptionBudget for capi services
+	capxPDB, err := getManifest("common", "capi_pdb.tmpl", keosCluster.Spec)
+	if err != nil {
+		return errors.Wrap(err, "failed to get PodDisruptionBudget file")
+	}
+	c = "echo '" + capxPDB + "' > " + capiPDBPath
 	_, err = commons.ExecuteCommand(n, c)
 	if err != nil {
-		return errors.Wrap(err, "failed to apply CAPX's NetworkPolicy in workload cluster")
+		return errors.Wrap(err, "failed to create PodDisruptionBudget file")
+	}
+
+	c = "kubectl --kubeconfig " + kubeconfigPath + " apply -f " + capiPDBPath
+	_, err = commons.ExecuteCommand(n, c)
+	if err != nil {
+		return errors.Wrap(err, "failed to apply "+p.capxName+" PodDisruptionBudget")
+	}
+
+	// Allow egress in CAPI's Namespaces
+	for _, deployment := range capiDeployments {
+		if !p.capxManaged || (p.capxManaged && !allowedNamePattern.MatchString(deployment.name)) {
+			c = "kubectl --kubeconfig " + kubeconfigPath + " -n " + deployment.namespace + " apply -f " + allowCommonEgressNetPolPath
+			_, err = commons.ExecuteCommand(n, c)
+			if err != nil {
+				return errors.Wrap(err, "failed to apply CAPI's egress NetworkPolicy in namespace "+deployment.namespace)
+			}
+		}
+	}
+
+	// Allow egress in cert-manager Namespace
+	c = "kubectl --kubeconfig " + kubeconfigPath + " -n cert-manager apply -f " + allowCommonEgressNetPolPath
+	_, err = commons.ExecuteCommand(n, c)
+	if err != nil {
+		return errors.Wrap(err, "failed to apply cert-manager's NetworkPolicy")
 	}
 
 	return nil