+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 0.17.0-0.3.4 (upcoming)
 
 * [Core] Conditionally increase replicas for capi controller manager services
+* [Core] Add PDB and PriorityClass to capx components
+
 
 ## 0.17.0-0.3.3 (2023-10-11)